Refactoring code to share between both extension tests
